Refactor code and repository structure to make it easier to add new features and maintain the codebase.
