+++ Makefile
@@ -1,0 +1,18 @@
+CC = g++
+
+SRC = src
+BIN = bin
+CFLAGS = -Wall -g -std=c++11 -lncurses
+
+all: pre-build glitch
+
+pre-build:
+	mkdir -p $(BIN)
+
+glitch: $(SRC)/glitch.cpp
+	$(CC) -o $(BIN)/glitch $(SRC)/glitch.cpp $(CFLAGS)
+
+clean:
+	rm -f glitch
+
+.PHONY: all clean
