+++ README.md
@@ -2,15 +2,27 @@
 git repo: https://github.com/NicholasNDev/glitch-effect
 simple command line glitch effect based on ncurses
 
-#Dependencieis
+## Dependencieis
 You will need ncurses library to compile this program. 
-You can install ncurses using your package manager as 'ncurses-dev' or 'libncurses-dev'.
+You can install ncurses using your package manager as `ncurses-dev` or `libncurses-dev`.
 
-#Compiling and running
-Compile the c++ script with g++: 'g++ -o glitch glitch.cpp -lncurses'.
-You can run your executable with './glitch'.
-For options run './glitch --help'.
+## Compiling and running
+First compile the c++ script with g++:
 
-#Where to start?
-You can move or copy your executable to /usr/bin so you can run it whenever you want.
-You need to specify a text file to display, you can start with files provided in my repo ^^^ 
+```
+g++ -o glitch glitch.cpp -lncurses
+```
+
+or use the makefile provided in the repo:
+
+```
+make
+```
+
+You can run your executable with `./glitch`.
+For options run `./glitch --help`.
+
+## Where to start?
+
+You can move or copy your executable to `/usr/bin` so you can run it whenever you want.
+You need to specify a text file to display, you can start with files provided in my repo.
